% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editing_exif.R
\name{editing_exif}
\alias{editing_exif}
\title{Batch editing of exif data}
\usage{
editing_exif(files, metadata)
}
\arguments{
\item{files}{a character vector of paths with complete file.names like what
\code{\link{list.files}} or
\code{\link[tools:fileutils]{tools::list_files_with_exts}} gives when
used with `full.names = TRUE`}

\item{metadata}{a data.frame as provided by \code{\link{parsing_nossaflex}}.}
}
\description{
Batch editing of exif data
}
\examples{
\dontrun{
files = tools::list_files_with_exts(dir = "tests/testthat/testdata/",
                    exts = "jpg", full.names = TRUE)

}

}
