% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_missing_records.R
\name{insert_missing_records}
\alias{insert_missing_records}
\title{Insert rows in metadata data.tables}
\usage{
insert_missing_records(metadata, i, value, extrapolate_data = FALSE)
}
\arguments{
\item{metadata}{a data.frame as provided by \code{\link{parsing_nossaflex}},
\code{\link{parsing_custom}} or \code{\link{parsing_json}}.}

\item{i}{An integer or integer vectors of row numbers where new values should
be inserted.}

\item{value}{A data.frame or data.table with column names matching names in
`metadata` and values to insert, probably Roll_Name and Camera_Brand for
instance.}

\item{extrapolate_data}{Logical, FALSE by default. If TRUE, the function
computes average Date_Time_Original, Latitude and Longitude values. Hence
works only with JSON data.}
}
\description{
Insert rows in metadata data.tables
}
\examples{
\dontrun{
# NOSSAFLEX data
metadata <- reading_nossaflex("tests/testthat/testdata/nossaflex_filenames.txt") |>
    parsing_nossaflex()
value <- data.table::data.table(T = "est roll", NO = "02", SS = "125",
                                A = "1.4", FL = "50", EX = "0")
insert_missing_records(metadata, 2, value)
# JSON data
metadata <- parsing_json("tests/testthat/testdata/nossaflex_Vito_70_test-week.txt")
value <- data.table::data.table(Roll_Name = rep("Test-week", 3),
                                Roll_Number = rep("005", 3),
                                Camera_Brand = rep("Voigtlaender", 3),
                                Camera_Model = rep("Vito 70", 3), NO = 5:7,
                                SS = rep("auto", 3), A = rep("auto", 3))
insert_missing_records(metadata = metadata, i = 5:7, value = value,
                       extrapolate_data = TRUE)
}
}
