% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nossaflex-package.R
\docType{package}
\name{nossaflex-package}
\alias{nossaflex}
\alias{nossaflex-package}
\title{nossaflex: Edit Exif Data of Pictures}
\description{
The nossaflex package is developed in GitHub
(https://github.com/albansagouis/nossaflex). To see the preferable citation of
the package, type citation("nossaflex").
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nossaflex.io/}
  \item \url{https://github.com/albansagouis/nossaflex/}
  \item Report bugs at \url{https://github.com/albansagouis/nossaflex/issues}
}

}
\keyword{internal}
