% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{parsing_custom}
\alias{parsing_custom}
\title{Parsing a custom NOSSAFLEX format
Parsing `filenames` with a consistent structure different from the classic
`NOSSAFLEX` structure into a data.frame}
\usage{
parsing_custom(filenames, format = "NO\%NO_SS\%SS_A\%A_FL\%FL_EX\%EX")
}
\arguments{
\item{filenames}{a character vector of file names following the structure:
NO01_SS125_A5.6_FL35_EX0
where
NO = Number
SS = Shutter Speed
A = Aperture
FL = Focal Length
EX = Exposure (Over/under)}

\item{format}{A character string. Used to extract the desired parts of the
NOSSAFLEX file names. default is "NO%NO_SS%SS_A%A_FL%FL_EX%EX" which
corresponds to the standard NOSSAFLEX structure as in
`NO01_SS125_A5.6_FL35_EX+1`.}
}
\description{
Parsing a custom NOSSAFLEX format
Parsing `filenames` with a consistent structure different from the classic
`NOSSAFLEX` structure into a data.frame
}
\details{
Inspired by \code{\link[base]{strptime}}
}
\keyword{internal}
