% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{parsing_frames}
\alias{parsing_frames}
\title{Parsing the JSON data sent by the `Frames` app}
\usage{
parsing_frames(path)
}
\arguments{
\item{path}{Complete path to the .frames / JSON data}
}
\value{
A data.frame with standard columns from nossaflex convention.
}
\description{
Parsing the JSON data sent by the `Frames` app
}
\examples{
path = '~/Projects/my r packages/nossaflex demo/Berlin_Boat.frames'
parsing_frames(path)
}
