% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{parsing_json}
\alias{parsing_json}
\title{Parsing the JSON data sent by the `Analog` app}
\usage{
parsing_json(path, apply_corrections = TRUE)
}
\arguments{
\item{path}{Complete path to the JSON data saved in a text file.}

\item{apply_corrections}{Correct known impossible values:

 * If Camera Brand is Voigtlaender and Camera Model is Vito 70
   * Aperture or Shutter Speed different from "auto" will be turned into "auto".
   * Lens Focal Length will be changed to 70mm
 * If Lens has a Maximum Aperture of 1.4, a wider aperture such as 1 will be
 changed into 1.4
 * If Aperture uses "," as decimal separator, a "." is written instead.}
}
\value{
A data.frame with standard columns from nossaflex convention.
}
\description{
Parsing the JSON data sent by the `Analog` app
}
\examples{
path = "~/idiv/my r packages/nossaflex demo/test1.json"
}
