% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lens_info.R
\name{add_lens_data_nikon_AF_28_105_D}
\alias{add_lens_data_nikon_AF_28_105_D}
\alias{add_lens_data_nikon_AF_24_D}
\alias{add_lens_data_nikon_AF_50_D}
\alias{add_lens_data_nikon_AF_85}
\alias{add_lens_data_nikon_AF_105_D}
\alias{add_lens_data_sigma_AF_28_70_D}
\title{Nikon AF 28-105 D}
\usage{
add_lens_data_nikon_AF_28_105_D()

add_lens_data_nikon_AF_24_D()

add_lens_data_nikon_AF_50_D()

add_lens_data_nikon_AF_85()

add_lens_data_nikon_AF_105_D()

add_lens_data_sigma_AF_28_70_D()
}
\description{
Nikon AF 28-105 D

Nikon AF 24 D

Nikon AF 50 D

Nikon AF 85

Nikon AF 105 D

Sigma AF 28-70 f2.8
}
\details{
To help `exiftoolr` to guess the lend ID, we can provide a lens of favourite
   lenses in a config file:
   # ~/.ExifTool_config
   # A special 'Lenses' list can be defined to give priority to specific lenses
   # in the logic to determine a lens model for the Composite:LensID tag
   @Image::ExifTool::UserDefined::Lenses = (
       'AF Zoom-Nikkor 28-105mm f/3.5-4.5D IF',
       'AF Nikkor 50mm f/1.4D',
       'AF Nikkor 85mm f/1.8',
       'AF Nikkor 24mm f/2.8D'
   );

  Additional lenses can be added here, the information listed here can be
  found in the exif data of pictures taken with the additional lenses.

  `Nikon:LensType` = 2 for D lenses and 0 for non-D lenses
  Exact values for the fields such as `Nikon:MinFocalLength` were obtained
  by shooting the lens with a DSLR and looking at the EXIF data.
}
\seealso{
See <https://exiftool.org/TagNames/Nikon.html#LensID> for the list
   of standard lens names
}
