% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{parsing_nossaflex}
\alias{parsing_nossaflex}
\title{Parsing nossaflex names into a data.frame}
\usage{
parsing_nossaflex(filenames)
}
\arguments{
\item{filenames}{a character vector of file names following the structure:
NO01_SS125_A5.6_FL35_EX0
where
NO = Number
SS = Shutter Speed
A = Aperture
FL = Focal Length
EX = Exposure (Over/under)}
}
\value{
a data.frame with standard columns from nossaflex convention.
}
\description{
Parsing nossaflex names into a data.frame
}
\examples{
filenames = c("NO01_SS1s_A5.6_FL35_EX-2", "NO02_SS125_A5.6_FL35_EX+1")
parsing_nossaflex(filenames)

}
